restructure views and routes
