+++ src/index.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import './index.css'
 import reportWebVitals from './reportWebVitals'
-import RouteTree from './routes/RouteTree'
 import axiosConfig from './networking/axiosConfig'
+import RouteTree from './views/RouteTree'
 
 axiosConfig()
 ReactDOM.render(
